relay_ourlogs

Function calculate_size

pub fn calculate_size(log: &OurLog) -> u64

Calculates a canonical size of a log item.

This size is passed along from Relay until storage and will eventually be emitted as an outcome.

Simple data types have a fixed size assigned to them:

Boolean: 1 byte

Integer: 8 byte

Double: 8 Byte

Strings are counted by their byte (UTF-8 encoded) representation.

Complex types like objects and arrays are counted as the sum of all contained simple data types.

Considered for the size of a log are all attribute keys, values and the log message body. An empty log (no message, no attributes) is counted as 1 byte.

The byte size should only be calculated once before any processing, enrichment and other modifications done by Relay.