relay_ourlogs

Module otel_logs

Structs

LogRecord

A log record according to OpenTelemetry Log Data Model: https://github.com/open-telemetry/oteps/blob/main/text/logs/0097-log-data-model.md

LogsData

LogsData represents the logs data that can be stored in a persistent storage, OR can be embedded by other protocols that transfer OTLP logs data but do not implement the OTLP protocol.

ResourceLogs

A collection of ScopeLogs from a Resource.

ScopeLogs

A collection of Logs produced by a Scope.

Enums

LogRecordFlags

LogRecordFlags represents constants used to interpret the LogRecord.flags field, which is protobuf ‘fixed32’ type and is to be used as bit-fields. Each non-zero value defined in this enum is a bit-mask. To extract the bit-field, for example, use an expression like:

SeverityNumber

Possible values for LogRecord.SeverityNumber.